--- src/main/scala/com/cognite/sdk/scala/v1/instances/InstanceDeletionRequest.scala
@@ -1,45 +1,0 @@
-package com.cognite.sdk.scala.v1.instances
-
-import io.circe.generic.semiauto.deriveDecoder
-import io.circe.syntax.EncoderOps
-import io.circe.{Decoder, DecodingFailure, Encoder, HCursor}
-
-sealed trait InstanceDeletionRequest
-
-object InstanceDeletionRequest {
-  final case class NodeDeletionRequest(space: String, externalId: String)
-      extends InstanceDeletionRequest {
-    val `type`: InstanceType = InstanceType.Node
-  }
-
-  final case class EdgeDeletionRequest(space: String, externalId: String)
-      extends InstanceDeletionRequest {
-    val `type`: InstanceType = InstanceType.Edge
-  }
-
-  implicit val nodeDeletionRequestEncoder: Encoder[NodeDeletionRequest] =
-    Encoder.forProduct3("type", "space", "externalId")((e: NodeDeletionRequest) =>
-      (e.`type`, e.space, e.externalId)
-    )
-
-  implicit val edgeDeletionRequestEncoder: Encoder[EdgeDeletionRequest] =
-    Encoder.forProduct3("type", "space", "externalId")((e: EdgeDeletionRequest) =>
-      (e.`type`, e.space, e.externalId)
-    )
-
-  implicit val instanceDeletionRequestEncoder: Encoder[InstanceDeletionRequest] = Encoder.instance {
-    case e: NodeDeletionRequest => e.asJson
-    case e: EdgeDeletionRequest => e.asJson
-  }
-
-  implicit val nodeDeletionRequestDecoder: Decoder[NodeDeletionRequest] = deriveDecoder
-
-  implicit val edgeDeletionRequestDecoder: Decoder[EdgeDeletionRequest] = deriveDecoder
-
-  implicit val instanceDeletionRequestDecoder: Decoder[InstanceDeletionRequest] = (c: HCursor) =>
-    c.downField("type").as[InstanceType] match {
-      case Left(err) => Left[DecodingFailure, InstanceDeletionRequest](err)
-      case Right(InstanceType.Node) => Decoder[NodeDeletionRequest].apply(c)
-      case Right(InstanceType.Edge) => Decoder[EdgeDeletionRequest].apply(c)
-    }
-}
