+++ src/test/scala/com/cognite/sdk/scala/v1/instances/FilterDefinitionSerDeTest.scala
@@ -1,7 +1,8 @@
 package com.cognite.sdk.scala.v1.instances
 
-import com.cognite.sdk.scala.v1.containers.ContainerReference
-import com.cognite.sdk.scala.v1.views.ViewReference
+import com.cognite.sdk.scala.v1.fdm.containers.ContainerReference
+import com.cognite.sdk.scala.v1.fdm.instances.FilterValueDefinition
+import com.cognite.sdk.scala.v1.fdm.views.ViewReference
 import io.circe.Printer
 import io.circe.parser._
 import io.circe.syntax._
@@ -18,7 +19,7 @@ import org.scalatest.wordspec.AnyWordSpec
   )
 )
 class FilterDefinitionSerDeTest extends AnyWordSpec with Matchers {
-  import FilterDefinition._
+  import com.cognite.sdk.scala.v1.fdm.instances.FilterDefinition._
   implicit val nullDroppingPrinter: Printer = Printer.noSpaces.copy(dropNullValues = true)
 
   "FilterDefinition Ser/de" when {
